[CET-6660] Move filter to standalone file
